refactor: scss, 컴포넌트 코드 분리
