+++ components/ContentCard/ContentCard.module.scss
@@ -1,0 +1,21 @@
+.contentCard {
+  padding: 8px;
+  color: white;
+  text-align: center;
+  background: #222;
+  border-radius: 8px;
+  outline: none;
+  transition: transform 0.2s, box-shadow 0.2s;
+}
+
+.focused {
+  transform: scale(1.05);
+  box-shadow: 0 0 0 3px #00c8ff;
+}
+
+.contentCard img {
+  width: 100%;
+  height: 300px;
+  object-fit: cover;
+  border-radius: 4px;
+}
